Refactoring repeated code and ignoring environment settings files
