+++ app/controllers/concert_controller.rb
@@ -1,4 +1,6 @@
 class ConcertController < ApplicationController
+  include CreatePlaylistHelper
+
   before_filter :require_authorization
   before_filter { require_band_member(params[:band_id]) }
   
@@ -78,21 +80,8 @@ class ConcertController < ApplicationController
   end
 
   def generate_playlist
-    if !current_user.is_spotify_user
-      flash[:alert] = "You have to connect your account to a Spotify account!"
-      redirect_to edit_user_path
-    else
-      concert = Concert.find(params[:id])
-      playlist_name = concert.name
-      # Use a thread for that
-      spotify_auth_token = current_user.spotify_oauth_token
-      spotify_user_id = current_user.spotify_uri.split(":")[-1]
-      CreatePlaylistWorker.perform_async(concert.setlist.id, playlist_name, spotify_auth_token, spotify_user_id)
-
-      flash[:notice] = 'We are creating your playlist in the background. It should finish after a while, check back soon!'
-
-      redirect_to :action => 'show'
-    end
+    concert = Concert.find(params[:id])
+    generate_playlist_for_setlist(concert.setlist.id, concert.name)
   end
 
   def concert_params
